+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.4.0] - 2021-03-25
+
 ### Added
 
 - Private key discovery in a directory.
@@ -53,7 +55,8 @@
 - Untracking of files.
 - Tracking registry initialization.
 
-[unreleased]: https://github.com/slok/agebox/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/slok/agebox/compare/v0.4.0...HEAD
+[v0.4.0]: https://github.com/slok/agebox/compare/v0.3.0...v0.4.0
 [v0.3.0]: https://github.com/slok/agebox/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slok/agebox/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/slok/agebox/compare/v0.1.0...v0.1.1